+++ CHANGELOG.md
@@ -59,6 +59,10 @@
 - some updates to composer packages. Removed deprecations in order to prepare for Symfony 6 ! 2617 @peter.toennies
 - Improved scripts. repaired outdated notify. Made test rerun faster. Made clean more powerfull !2622 @peter.toennies
 
+
+- Delete faulty mails on imap server !2591 @stcz
+
+
 ## Development documentation